+++ srcs/raycast_calculs/misc_functs.c
@@ -55,7 +55,7 @@ int	check_wall_colision(t_fcoords pos, char **map_arr)
 	return (0);
 }
 /* -------------------------------------------------------------------------- */
-double dstnce_btwn_points(float x1, float y1, float x2, float y2)
+double dstnce_btwn_points(double x1, double y1, double x2, double y2)
 {
     return (sqrt((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1)));
 }
